feat: Implement PUT endpoint for updating payment status in pending_to_pay API
feat: Enhance invoices page with dynamic table, filtering, and Excel export functionality

refactor: Remove create order page and clean up related code

fix: Update orders page to remove unused icons and adjust action buttons

style: Revamp finance page layout and navigation for better user experience

chore: Delete unused pending-to-pay pages and components

feat: Add invoice display page for detailed invoice view

fix: Update sidebar icons and links for consistency in finance module
