+++ app/api/finance/invoices/route.ts
@@ -11,3 +11,30 @@ export async function GET() {
   
   return NextResponse.json(quotations);
 }
+
+export async function PUT(request: Request) {
+  const supabase = await createClient();
+  const body = await request.json();
+  const {id, status} = body;
+
+  if(!id){
+    return NextResponse.json({ error: "Invoice ID is required" }, { status: 400 });
+  }
+  
+  const allowedStatuses = ["Issued", "Expired"];
+
+  if (status && !allowedStatuses.includes(status)) {
+    return NextResponse.json({ error: "Invalid invoice status" }, { status: 400 });
+  }
+
+  const { data, error } = await supabase.from("invoices").update({status}).eq("invoice_id", body.id);
+
+  if (error) {
+    return NextResponse.json({ error: error.message }, { status: 500 });
+  }
+
+  return NextResponse.json({
+    message: "Invoice updated successfully",
+    data,
+   });
+}
