+++ app/api/finance/orders/route.ts
@@ -11,3 +11,30 @@ export async function GET() {
 
   return NextResponse.json(orders);
 }
+
+export async function PUT(request: Request) {
+  const supabase = await createClient();
+  const body = await request.json();
+  const {id, status} = body;
+
+  if(!id){
+    return NextResponse.json({ error: "Order ID is required" }, { status: 400 });
+  }
+  
+  const allowedStatuses = ["Paid", "Cancelled"];
+
+  if (status && !allowedStatuses.includes(status)) {
+    return NextResponse.json({ error: "Invalid order status" }, { status: 400 });
+  }
+
+  const { data, error } = await supabase.from("orders").update({status}).eq("order_id", body.id);
+
+  if (error) {
+    return NextResponse.json({ error: error.message }, { status: 500 });
+  }
+
+  return NextResponse.json({
+    message: "Order updated successfully",
+    data,
+   });
+}
