+++ app/api/finance/pending_to_pay/route.ts
@@ -11,3 +11,30 @@ export async function GET() {
 
   return NextResponse.json(pending_to_pay);
 }
+
+export async function PUT(request: Request) {
+  const supabase = await createClient();
+  const body = await request.json();
+  const {id, status} = body;
+
+  if(!id){
+    return NextResponse.json({ error: "Payable ID is required" }, { status: 400 });
+  }
+  
+  const allowedStatuses = ["Paid", "Expired"];
+
+  if (status && !allowedStatuses.includes(status)) {
+    return NextResponse.json({ error: "Invalid payable status" }, { status: 400 });
+  }
+
+  const { data, error } = await supabase.from("pending_to_pay").update({status}).eq("payable_id", body.id);
+
+  if (error) {
+    return NextResponse.json({ error: error.message }, { status: 500 });
+  }
+
+  return NextResponse.json({
+    message: "Payment updated successfully",
+    data,
+   });
+}
