--- app/finance/pending-to-pay/[id]/page.tsx
@@ -1,8 +1,0 @@
-export default function EditPage() {
-  return (
-    <div>
-      <h1>view id </h1>
-      {/* Render your invoices here */}
-    </div>
-  );
-}
