+++ components/DynamicTable.tsx
@@ -55,6 +55,11 @@ type Props = {
     icon2?: React.ReactNode;
     icon3?: React.ReactNode;
   };
+  actions?: {
+    view?: boolean;
+    accept?: boolean;
+    cancel?: boolean;
+  };
 };
 
 export default function DynamicTable({
@@ -65,6 +70,11 @@ export default function DynamicTable({
   onDataChange,
   actionHandlers,
   actionIcons,
+  actions = {
+    view: true,
+    accept: true,
+    cancel: true,
+  },
 }: Props) {
   const [selectedRows, setSelectedRows] = useState<string[]>([]);
   const [isActive, setIsActive] = useState<boolean | undefined>();
@@ -81,45 +91,42 @@ export default function DynamicTable({
   const paginatedData = data.slice(startIndex, startIndex + rowsPerPage);
 
   useEffect(() => {
-  setData(initialData);
-}, [initialData]);
+    setData(initialData);
+  }, [initialData]);
 
   useEffect(() => {
-  setData(initialData);
-}, [initialData]);
+    setData(initialData);
+  }, [initialData]);
 
-useEffect(() => {
-  console.log("el estado es: "+isActive);
-},[isActive]);
-  
+  useEffect(() => {
+    console.log("el estado es: " + isActive);
+  }, [isActive]);
+
   const toggleRow = (rowId: string) => {
-    const updated = selectedRows.includes(rowId)
-      ? selectedRows.filter((id) => id !== rowId)
-      : [...selectedRows, rowId];
+    const updated = selectedRows.includes(rowId) ? selectedRows.filter((id) => id !== rowId) : [...selectedRows, rowId];
 
     setSelectedRows(updated);
     if (onSelectedRowsChange) onSelectedRowsChange(updated);
   };
 
   const toggleActive = (rowId: string) => {
-  const newData = data.map((row) => {
-    if (row.id === rowId) {
-      const newActive = !row.active;
-      setIsActive(newActive);
-      return { ...row, active: newActive };
+    const newData = data.map((row) => {
+      if (row.id === rowId) {
+        const newActive = !row.active;
+        setIsActive(newActive);
+        return { ...row, active: newActive };
+      }
+      return row;
+    });
+
+    setData(newData);
+    if (onDataChange) onDataChange(newData);
+
+    const changedRow = newData.find((row) => row.id === rowId);
+    if (onActiveChange && changedRow) {
+      onActiveChange([{ id: rowId, active: changedRow.active as boolean }]);
     }
-    return row;
-  });
-
-  setData(newData);
-  if (onDataChange) onDataChange(newData);
-
-  const changedRow = newData.find((row) => row.id === rowId);
-  if (onActiveChange && changedRow) {
-    onActiveChange([{ id: rowId, active: changedRow.active as boolean }]);
-  }
-};
-
+  };
 
   if (!data || data.length === 0) {
     return (
@@ -128,10 +135,7 @@ useEffect(() => {
           <thead>
             <tr className="bg-[#a01217] text-white">
               {columns.map((col) => (
-                <th
-                  key={col.key}
-                  className="px-4 py-2 text-left border-b border-black"
-                >
+                <th key={col.key} className="px-4 py-2 text-left border-b border-black">
                   {col.label}
                 </th>
               ))}
@@ -151,10 +155,7 @@ useEffect(() => {
         <thead>
           <tr className="bg-[#a01217] text-white">
             {columns.map((col) => (
-              <th
-                key={col.key}
-                className="px-4 py-2 text-left border-b border-black"
-              >
+              <th key={col.key} className="px-4 py-2 text-left border-b border-black">
                 {col.label}
               </th>
             ))}
@@ -165,57 +166,46 @@ useEffect(() => {
             const isSelected = selectedRows.includes(row.id);
 
             return (
-              <tr
-                key={idx}
-                className={`border-b border-black transition-all ${
-                  isSelected ? "bg-black/10" : "hover:bg-black/10"
-                }`}
-              >
+              <tr key={idx} className={`border-b border-black transition-all ${isSelected ? "bg-black/10" : "hover:bg-black/10"}`}>
                 {columns.map((col) => (
                   <td key={col.key} className="px-4 py-2">
                     {col.type === "checkbox" ? (
-                      <input
-                        type="checkbox"
-                        checked={isSelected}
-                        onChange={() => toggleRow(row.id)}
-                        className="w-5 h-5 accent-red-600"
-                      />
+                      <input type="checkbox" checked={isSelected} onChange={() => toggleRow(row.id)} className="w-5 h-5 accent-red-600" />
                     ) : col.type === "switch" ? (
                       <label className="relative inline-flex items-center cursor-pointer">
-                        <input
-                          type="checkbox"
-                          checked={!!row.active}
-                          onChange={() => toggleActive(row.id)}
-                          className="sr-only peer"
-                        />
+                        <input type="checkbox" checked={!!row.active} onChange={() => toggleActive(row.id)} className="sr-only peer" />
                         <div className="w-11 h-6 bg-gray-400 peer-focus:outline-none rounded-full peer peer-checked:after:translate-x-full after:content-[''] after:absolute after:top-0.5 after:left-[4px] after:bg-black after:border after:rounded-full after:h-5 after:w-5 after:transition-transform peer-checked:bg-[#a01217]"></div>
                       </label>
                     ) : col.type === "action" ? (
+
                       <div className="flex gap-2">
-                        <button
-                          onClick={() =>
-                            actionHandlers?.onView?.(row.id)
-                          }
-                          className="p-2 text-[#a01217] bg-[#a0121722] rounded-full hover:bg-[#a0121744]"
-                        >
-                          {actionIcons?.icon1 ?? <FaEye className="w-5 h-5" />}
-                        </button>
-                        <button
-                          onClick={() =>
-                            actionHandlers?.onAccept?.(row.id) 
-                          }
-                          className="p-2 text-[#a01217] bg-[#a0121722] rounded-full hover:bg-[#a0121744]"
-                        >
-                          {actionIcons?.icon2 ?? <FaEdit className="w-5 h-5" />}
-                        </button>
-                        <button
-                          onClick={() =>
-                            actionHandlers?.onCancel?.(row.id) 
-                          }
-                          className="p-2 text-[#a01217] bg-[#a0121722] rounded-full hover:bg-[#a0121744]"
-                        >
-                          {actionIcons?.icon3 ?? <FaTrash className="w-5 h-5" />}
-                        </button>
+
+                        {actions.view && (
+                          <button
+                            onClick={() => actionHandlers?.onView?.(row.id)}
+                            className="p-2 text-[#a01217] bg-[#a0121722] rounded-full hover:bg-[#a0121744]"
+                          >
+                            {actionIcons?.icon1 ?? <FaEye className="w-5 h-5" />}
+                          </button>
+                        )}
+
+                        {actions.accept && (
+                          <button
+                            onClick={() => actionHandlers?.onEdit?.(row.id)}
+                            className="p-2 text-[#a01217] bg-[#a0121722] rounded-full hover:bg-[#a0121744]"
+                          >
+                            {actionIcons?.icon2 ?? <FaEdit className="w-5 h-5" />}
+                          </button>
+                        )}
+
+                        {actions.cancel && (
+                          <button
+                            onClick={() => actionHandlers?.onCancel?.(row.id)}
+                            className="p-2 text-[#a01217] bg-[#a0121722] rounded-full hover:bg-[#a0121744]"
+                          >
+                            {actionIcons?.icon3 ?? <FaTrash className="w-5 h-5" />}
+                          </button>
+                        )}
                       </div>
                     ) : (
                       row[col.key]
@@ -228,29 +218,28 @@ useEffect(() => {
         </tbody>
       </table>
 
-{/* Pagination controls */}
-<div className="flex items-center justify-center mt-4 gap-4">
-  <button
-    onClick={() => setCurrentPage((prev) => Math.max(prev - 1, 1))}
-    disabled={currentPage === 1}
-    className={`px-4 py-2 rounded bg-[#a01217] text-white hover:bg-[#7f0e12] disabled:bg-gray-300`}
-  >
-    ‹
-  </button>
+      {/* Pagination controls */}
+      <div className="flex items-center justify-center mt-4 gap-4">
+        <button
+          onClick={() => setCurrentPage((prev) => Math.max(prev - 1, 1))}
+          disabled={currentPage === 1}
+          className={`px-4 py-2 rounded bg-[#a01217] text-white hover:bg-[#7f0e12] disabled:bg-gray-300`}
+        >
+          ‹
+        </button>
 
-  <span className="text-black">
-    Page {currentPage} of {totalPages}
-  </span>
+        <span className="text-black">
+          Page {currentPage} of {totalPages}
+        </span>
 
-  <button
-    onClick={() => setCurrentPage((prev) => Math.min(prev + 1, totalPages))}
-    disabled={currentPage === totalPages}
-    className={`px-4 py-2 rounded bg-[#a01217] text-white hover:bg-[#7f0e12] disabled:bg-gray-300`}
-  >
-    ›
-  </button>
-</div>
-
+        <button
+          onClick={() => setCurrentPage((prev) => Math.min(prev + 1, totalPages))}
+          disabled={currentPage === totalPages}
+          className={`px-4 py-2 rounded bg-[#a01217] text-white hover:bg-[#7f0e12] disabled:bg-gray-300`}
+        >
+          ›
+        </button>
+      </div>
     </div>
   );
 }
