+++ components/Sidebar.tsx
@@ -15,7 +15,7 @@ import {
   FiUserCheck,
 } from 'react-icons/fi';
 import { FaHandHoldingUsd } from "react-icons/fa";
-import { /*FaFileInvoiceDollar,*/FaClipboardCheck } from "react-icons/fa6";
+import { FaClipboardList, FaFileInvoiceDollar } from "react-icons/fa6";
 import { BsHouse } from 'react-icons/bs';
 import Image from 'next/image';
 import styles from '../styles/Sidebar.module.css';
@@ -35,8 +35,8 @@ export default function Sidebar() {
     { name: 'Finance',
       icon: <FiDollarSign />, 
       submodules: [
-        { name: 'Orders', href: '/finance/orders', icon: <FaClipboardCheck /> },
-        //{ name: 'Invoices', href: '/finance/invoices', icon: <FaFileInvoiceDollar /> },
+        { name: 'Orders', href: '/finance/orders', icon: <FaClipboardList /> },
+        { name: 'Invoices', href: '/finance/invoices', icon: <FaFileInvoiceDollar /> },
         { name: 'Pending to pay', href: '/finance/pending-to-pay', icon: <FaHandHoldingUsd /> },
       ],
     },
